+++ examples/updates/group/main.tf
@@ -1,0 +1,9 @@
+variable "description" {
+  type = string
+}
+
+resource "genesyscloud_group" "eng" {
+  name        = "Engineering-updates"
+  type        = "official"
+  description = var.description
+}