+++ examples/updates/location/main.tf
@@ -1,0 +1,8 @@
+variable "notes" {
+  type = string
+}
+
+resource "genesyscloud_location" "office" {
+  name  = "Office-updates"
+  notes = var.notes
+}