+++ examples/updates/oauth_client/main.tf
@@ -1,0 +1,9 @@
+variable "description" {
+  type = string
+}
+
+resource "genesyscloud_oauth_client" "bot" {
+  name                  = "bot-updates"
+  description           = var.description
+  authorized_grant_type = "CLIENT_CREDENTIALS"
+}