+++ examples/updates/user/main.tf
@@ -1,0 +1,9 @@
+variable "title" {
+  type = string
+}
+
+resource "genesyscloud_user" "alice" {
+  name  = "Alice Example"
+  email = "alice-updates@example.com"
+  title = var.title
+}