+++ examples/working/group/main.tf
@@ -1,0 +1,5 @@
+resource "genesyscloud_group" "engineering" {
+  name        = "Engineering"
+  type        = "official"
+  description = "Engineering group"
+}