+++ examples/working/location/main.tf
@@ -1,0 +1,3 @@
+resource "genesyscloud_location" "hq" {
+  name = "Headquarters"
+}