+++ examples/working/oauth_client/main.tf
@@ -1,0 +1,4 @@
+resource "genesyscloud_oauth_client" "integration_bot" {
+  name                  = "integration-bot"
+  authorized_grant_type = "CLIENT_CREDENTIALS"
+}